Tidy namespaces.
