+++ osu.Game/Beatmaps/Drawable/BeatmapGroup.cs
@@ -2,24 +2,18 @@
 //Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
 using System;
+using System.Linq;
+using osu.Framework;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Sprites;
-using osu.Game.Graphics;
-using osu.Game.Beatmaps;
-using osu.Game.Database;
 using osu.Framework.Graphics.Primitives;
-using OpenTK;
-using System.Linq;
-using osu.Framework.Graphics.Transformations;
 using osu.Framework.Input;
+using osu.Game.Database;
+using osu.Game.Graphics;
+using OpenTK;
 using OpenTK.Graphics;
-using osu.Game.Beatmaps.IO;
-using osu.Framework.Graphics.Textures;
-using System.Threading.Tasks;
-using osu.Framework;
 
-namespace osu.Game.GameModes.Play
+namespace osu.Game.Beatmaps.Drawable
 {
     class BeatmapGroup : Container, IStateful<BeatmapGroup.GroupState>
     {
@@ -44,7 +38,7 @@ namespace osu.Game.GameModes.Play
 
         public Action<BeatmapGroup, BeatmapInfo> BeatmapSelected;
         public BeatmapSetInfo BeatmapSet;
-        private BeatmapSetHeader setBox;
+        private BeatmapSetHeader header;
         private FlowContainer difficulties;
         private bool collapsed;
         public GroupState State
@@ -63,14 +57,14 @@ namespace osu.Game.GameModes.Play
                     difficulties.Hide();
                 else
                     difficulties.Show();
-                setBox.ClearTransformations();
-                setBox.Width = collapsed ? collapsedWidth : 1; // TODO: Transform
-                setBox.BorderColour = new Color4(
-                    setBox.BorderColour.R,
-                    setBox.BorderColour.G,
-                    setBox.BorderColour.B,
+                header.ClearTransformations();
+                header.Width = collapsed ? collapsedWidth : 1; // TODO: Transform
+                header.BorderColour = new Color4(
+                    header.BorderColour.R,
+                    header.BorderColour.G,
+                    header.BorderColour.B,
                     collapsed ? 0 : 255);
-                setBox.GlowRadius = collapsed ? 0 : 5;
+                header.GlowRadius = collapsed ? 0 : 5;
             }
         }
 
@@ -89,9 +83,9 @@ namespace osu.Game.GameModes.Play
                     RelativeSizeAxes = Axes.X,
                     AutoSizeAxes = Axes.Y,
                     Direction = FlowDirection.VerticalOnly,
-                    Children = new Drawable[]
+                    Children = new Framework.Graphics.Drawable[]
                     {
-                        setBox = new BeatmapSetHeader(beatmapSet)
+                        header = new BeatmapSetHeader(beatmapSet)
                         {
                             RelativeSizeAxes = Axes.X,
                             Width = collapsedWidth,
@@ -140,91 +134,6 @@ namespace osu.Game.GameModes.Play
         {
             BeatmapSelected?.Invoke(this, selectedBeatmap);
             return true;
-        }
-    }
-    
-    class BeatmapSetHeader : Container
-    {
-        public BeatmapSetHeader(BeatmapSetInfo beatmapSet)
-        {
-            AutoSizeAxes = Axes.Y;
-            Masking = true;
-            CornerRadius = 5;
-            BorderThickness = 2;
-            BorderColour = new Color4(221, 255, 255, 0);
-            GlowColour = new Color4(166, 221, 251, 0.5f); // TODO: Get actual color for this
-            Children = new Drawable[]
-            {
-                new Box
-                {
-                    Colour = new Color4(85, 85, 85, 255),
-                    RelativeSizeAxes = Axes.Both,
-                    Size = Vector2.One,
-                },
-                new Container
-                {
-                    RelativeSizeAxes = Axes.Both,
-                    Size = Vector2.One,
-                    Children = new Drawable[]
-                    {
-                        new Box // TODO: Gradient
-                        {
-                            Colour = new Color4(0, 0, 0, 100),
-                            RelativeSizeAxes = Axes.Both,
-                            Size = Vector2.One,
-                        }
-                    }
-                },
-                new FlowContainer
-                {
-                    Direction = FlowDirection.VerticalOnly,
-                    Spacing = new Vector2(0, 2),
-                    Padding = new MarginPadding { Top = 3, Left = 20, Right = 20, Bottom = 3 },
-                    AutoSizeAxes = Axes.Both,
-                    Children = new[]
-                    {
-                        // TODO: Make these italic
-                        new SpriteText
-                        {
-                            Text = beatmapSet.Metadata.Title ?? beatmapSet.Metadata.TitleUnicode,
-                            TextSize = 20
-                        },
-                        new SpriteText
-                        {
-                            Text = beatmapSet.Metadata.Artist ?? beatmapSet.Metadata.ArtistUnicode,
-                            TextSize = 16
-                        },
-                        new FlowContainer
-                        {
-                            AutoSizeAxes = Axes.Both,
-                            Children = new[]
-                            {
-                                new DifficultyIcon(FontAwesome.dot_circle_o, new Color4(159, 198, 0, 255)),
-                                new DifficultyIcon(FontAwesome.dot_circle_o, new Color4(246, 101, 166, 255)),
-                            }
-                        }
-                    }
-                }
-            };
-        }
-    }
-    
-    class DifficultyIcon : Container
-    {
-        public DifficultyIcon(FontAwesome icon, Color4 color)
-        {
-            const float size = 20;
-            Size = new Vector2(size);
-            Children = new[]
-            {
-                new TextAwesome
-                {
-                    Anchor = Anchor.Centre,
-                    TextSize = size,
-                    Colour = color,
-                    Icon = icon
-                }
-            };
         }
     }
 }
